api/requirements.txt
pinecone-client==2.2.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
Flask==3.0.0
Jinja2==3.1.2
Werkzeug==3.0.1
Flask-Cors==4.0.0
